refactor: migrate arrays to objects for better UX
There are two reasons behind keeping the data sources as arrays,
to then convert them to plain objects.

First, in a real life situation, a platform might have a very large
amount of contracts/customers, which will be quite slow to update by using
arrays, if we are to look for them by id. Not only that, the methods to
update would be hard to read and add unneeded complexity while debugging.

Second, to show understanding of data handling as part of the coding
challenge, but also to show understanding of the load we put in the users
computer, when we execute large operations. Having the info as objects is
nice, but the migration process might take too long.